refactoring: replacing obsolete Math class
