+++ library/src/Core/Filtering/Components/ButterworthFilter.cs
@@ -1,6 +1,6 @@
 ﻿
-using System;
 using ReFlex.Core.Common.Components;
+using Math = System.Math;
 
 namespace ReFlex.Core.Filtering.Components
 {
